pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gitpython>=3.1.45",
    "gunicorn>=23.0.0",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "python-telegram-bot>=22.3",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.42",
    "stripe>=12.4.0",
    "telegram>=0.0.1",
    "werkzeug>=3.1.3",
]
